pyproject.toml
[tool.poetry]
name = "e-commerce fastapi"
version = "0.1.0"
description = ""
authors = ["Ifeoluwaobayemi <“ifeoluwaobayemi00@gmail.com”>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
pydantic = {extras = ["email-validator"], version = "^2.9.2"}
uvicorn = "^0.31.0"
sqlalchemy = "^2.0.35"
asyncpg = "^0.29.0"
alembic = "^1.13.3"
python-jose = "^3.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-dotenv = "^1.0.1"
black = "^24.10.0"
flake8 = "^7.1.1"
pre-commit = "^4.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
